Changed piece parents. General selection scripts.
Added the selection scripts from the Pawn to the Rook and Bishop.

Changed the names of the scripts to make it clear what each script does.

Moved the pieces from the main scene to the board scene. Each piece is now a child of the tile it is on, which should make it easier to detect empty tiles.

I also made it so two pieces cannot occupy the same tile, but this might need to be changed in the future.