Added Colors and Adjusted Piece Movement.
Added color to both the pieces and the board. Requires that Tile and Piece names not be changed.

Fixed movement: Piece instances themselves are now parented to the tile instance itself. Previously, the contents of piece instances were parented to the contents of the tile instance. With this change, the names of the pieces and the tiles can be referenced.